added Azure SDK for DNS
