+++ federation/pkg/dnsprovider/providers/azure/azuredns/helpers.go
@@ -17,22 +17,9 @@ limitations under the License.
 package azuredns
 
 import (
-	"encoding/json"
-
 	"github.com/Azure/go-autorest/autorest/adal"
 	"github.com/Azure/go-autorest/autorest/azure"
 )
-
-const (
-	credentialsPath = "/.azure/credentials.json"
-)
-
-// ToJSON returns the passed item as a pretty-printed JSON string. If any JSON error occurs,
-// it returns the empty string.
-func ToJSON(v interface{}) (string, error) {
-	j, err := json.MarshalIndent(v, "", "  ")
-	return string(j), err
-}
 
 // NewServicePrincipalTokenFromCredentials creates a new ServicePrincipalToken using values of the
 // passed credentials map.
